Implemented custom classes for some event listeners
