+++ src/me/mrCookieSlime/Slimefun/listeners/JoinGuideListener.java
@@ -1,0 +1,30 @@
+package me.mrCookieSlime.Slimefun.listeners;
+
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
+
+import me.mrCookieSlime.Slimefun.SlimefunGuide;
+import me.mrCookieSlime.Slimefun.SlimefunStartup;
+import me.mrCookieSlime.Slimefun.Misc.BookDesign;
+
+public class JoinGuideListener implements Listener {
+
+	public JoinGuideListener(SlimefunStartup plugin) {
+		plugin.getServer().getPluginManager().registerEvents(this, plugin);
+	}
+
+	@EventHandler
+	public void onJoin(PlayerJoinEvent e) {
+		if (!e.getPlayer().hasPlayedBefore()) {
+			Player p = e.getPlayer();
+			if (!SlimefunStartup.getWhitelist().getBoolean(p.getWorld().getName() + ".enabled")) return;
+			if (!SlimefunStartup.getWhitelist().getBoolean(p.getWorld().getName() + ".enabled-items.SLIMEFUN_GUIDE")) return;
+			
+			p.getInventory().addItem(SlimefunStartup.getCfg().getBoolean("guide.default-view-book")
+					? SlimefunGuide.getItem(BookDesign.BOOK) : SlimefunGuide.getItem(BookDesign.CHEST));
+		}
+	}
+
+}
